+++ src/extensions/techno/technoext_hooks.cpp
@@ -69,6 +69,7 @@
 #include "textprint.h"
 #include "tiberiumext.h"
 #include "unittype.h"
+#include "unittypeext.h"
 #include "verses.h"
 #include "session.h"
 #include "mouse.h"
@@ -1579,6 +1580,58 @@ DECLARE_PATCH(_TechnoClass_Take_Damage_Drop_Tiberium_Type_Patch)
 
 
 /**
+ *  #issue-715
+ *
+ *  Enables the deploy keyboard command to work for units that
+ *  transform into a different unit on deploy.
+ *
+ *  @author: Rampastring
+ */
+DECLARE_PATCH(_TechnoClass_2A0_Is_Allowed_To_Deploy_Unit_Transform_Patch)
+{
+    GET_REGISTER_STATIC(UnitTypeClass*, unittype, eax);
+    static UnitTypeClassExtension* unittypeext;
+
+    /**
+     *  Stolen bytes/code.
+     */
+    if (unittype->DeploysInto != nullptr) {
+        goto has_deploy_ability;
+
+    }
+    else if (unittype->MaxPassengers > 0) {
+        goto has_deploy_ability;
+
+    }
+    else if (unittype->IsMobileEMP) {
+        goto has_deploy_ability;
+    }
+
+    unittypeext = Extension::Fetch<UnitTypeClassExtension>(unittype);
+
+    if (unittypeext->TransformsInto != nullptr) {
+        goto has_deploy_ability;
+    }
+
+    /**
+     *  The unit has no ability that allows it to deploy / unload.
+     *  Mark that and continue function after the check.
+     */
+has_no_deploy_ability:
+    _asm { mov eax, unittype }
+    JMP_REG(ecx, 0x006320E0);
+
+    /**
+     *  The unit has some kind of an ability that allows it to deploy / unload.
+     *  Continue function after the check.
+     */
+has_deploy_ability:
+    _asm { mov eax, unittype }
+    JMP_REG(ecx, 0x006320E5);
+}
+
+
+/**
  *  Main function for patching the hooks.
  */
 void TechnoClassExtension_Hooks()
@@ -1610,4 +1663,5 @@ void TechnoClassExtension_Hooks()
     Patch_Call(0x004A8532, &TechnoClassExt::_What_Action);
     Patch_Jump(0x0062EB27, &_TechnoClass_AI_Abandon_Invalid_Target_Patch);
     Patch_Jump(0x00632F4C, &_TechnoClass_Take_Damage_Drop_Tiberium_Type_Patch);
+    Patch_Jump(0x006320C2, &_TechnoClass_2A0_Is_Allowed_To_Deploy_Unit_Transform_Patch);
 }